AccountsAndUsers/src/main/java/ru/itterminal/botdesk/aau/service/validator/GroupOperationValidator.java
package ru.itterminal.botdesk.aau.service.validator;

import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;
import ru.itterminal.botdesk.aau.model.Group;
import ru.itterminal.botdesk.aau.model.Role;
import ru.itterminal.botdesk.aau.model.projection.GroupUniqueFields;
import ru.itterminal.botdesk.aau.model.projection.UserUniqueFields;
import ru.itterminal.botdesk.aau.repository.RoleRepository;
import ru.itterminal.botdesk.aau.service.impl.GroupServiceImpl;
import ru.itterminal.botdesk.aau.service.impl.UserServiceImpl;
import ru.itterminal.botdesk.commons.exception.LogicalValidationException;
import ru.itterminal.botdesk.commons.exception.error.ValidationError;
import ru.itterminal.botdesk.commons.service.validator.impl.BasicOperationValidatorImpl;

import java.util.HashMap;
import java.util.List;
import java.util.Map;

import static java.lang.String.format;
import static java.util.Collections.singletonList;

@Slf4j
@Component
public class GroupOperationValidator extends BasicOperationValidatorImpl<Group> {
    private GroupServiceImpl service;
    private RoleRepository repository;

    @Autowired
    public GroupOperationValidator(GroupServiceImpl service, RoleRepository repository) {
        this.service = service;
        this.repository = repository;
    }

    @Override
    public boolean beforeCreate(Group entity) {
        super.beforeCreate(entity);
        checkUniqueness(entity);
        return true;
    }

    @Override
    public boolean beforeUpdate(Group entity) {
        super.beforeCreate(entity);
        checkUniqueness(entity);
        return true;
    }

    @Override
    public boolean checkUniqueness(Group entity) {
        log.trace(CHECK_UNIQUENESS, entity);
        Map<String, List<ValidationError>> errors = new HashMap<>();
        List<GroupUniqueFields> foundGroup = service.findByUniqueFields(entity);
        if (foundGroup.isEmpty()) {
            log.trace(FIELDS_UNIQUE, entity);
            return true;
        } else {
            String validatedField;
            if (entity.getName().equalsIgnoreCase(foundGroup.get(0).getName())) {
                validatedField = "name";
                errors.put(validatedField, singletonList(new ValidationError(NOT_UNIQUE_CODE,
                        format(NOT_UNIQUE_MESSAGE, validatedField))));
            }
            log.error(FIELDS_NOT_UNIQUE, errors);
            throw new LogicalValidationException(VALIDATION_FAILED, errors);
        }
    }
}
